map form works, as does geocoding

+++ Location.php
@@ -10,8 +10,8 @@ class Location extends Kea_Record
 		$this->setTableName('locations');
 		
 		$this->hasColumn('item_id', 'integer', null, array('range'=>array('1')));
-		$this->hasColumn('latitude', 'double', null, array('notblank'=>true));
-		$this->hasColumn('longitude', 'double', null, array('notblank'=>true));
+		$this->hasColumn('latitude', 'float', null, array('notblank'=>true));
+		$this->hasColumn('longitude', 'float', null, array('notblank'=>true));
 		$this->hasColumn('zipcode', 'integer');
 		$this->hasColumn('zoom_level', 'integer');
 		$this->hasColumn('map_type', 'string', 255);
